+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0
+### Added
+- Compatibility with `Scrollable` widgets (such as `ListViews`).
+
 ## 1.1.1
 ### Fixed
 - Typo in Readme.
@@ -8,7 +12,7 @@
 
 ## 1.0.1
 ### Fixed
-- Fix debug properties in `Gap` widget
+- Fix debug properties in `Gap` widget.
 
 ## 1.0.0
-- Initial Open Source release
+- Initial Open Source release.